Fix structure after major rebase
